creates color swatch with sizing system
