+++ src/components/Display.svelte
@@ -8,6 +8,10 @@
 </script>
 
 <div class="flex flex-col w-full items-center justify-center flex-1">
-  <ColorTitle color={current} />
+  <ColorTitle
+    color={current}
+    swatchSize="md"
+    caption="The selected color is "
+  />
   <Square {squares} {handleSelectColor} />
 </div>
