+++ src/types/Color.ts
@@ -14,9 +14,3 @@ export interface ColorPropsOptions
 }
 
 export type ColorPropKey = keyof ColorProps
-
-export interface ControlOptions {
-  property: ColorPropKey
-  width: number
-  step: number
-}
