+++ src/types/ControlOptions.ts
@@ -1,0 +1,7 @@
+import type { ColorPropKey } from './Color'
+
+export interface ControlOptions {
+  property: ColorPropKey
+  width: number
+  step: number
+}
