+++ src/types/index.ts
@@ -1,3 +1,5 @@
 export * from './Color'
 export * from './Square'
 export * from './Svg'
+export * from './Swatch'
+export * from './ControlOptions'
